@@ -1,103 +1,39 @@
 package rw.ac.rca.WebsiteDownloadManager.server.models;
 
-import org.hibernate.annotations.GenericGenerator;
-import org.hibernate.annotations.Type;
+import com.fasterxml.jackson.annotation.JsonFormat;
+import com.fasterxml.jackson.annotation.JsonManagedReference;
+import lombok.Getter;
+import lombok.Setter;
 
 import javax.persistence.*;
-import java.util.Date;
+import java.time.LocalDateTime;
+import java.util.HashSet;
+import java.util.Set;
 import java.util.UUID;
 
+@Getter
+@Setter
 @Entity
-@Table(name = "website_tbl")
+@Table(name = "websites")
 public class Website {
-  @Id
-  @GeneratedValue(generator = "UUID")
-  @GenericGenerator(name = "UUID", strategy = "org.hibernate.id.UUIDGenerator")
-  @Type(type="uuid-char")
-  @Column(columnDefinition = "VARCHAR(255)")
-  private UUID id;
-  @Column(columnDefinition = "TEXT",name = "website_name" )
-  private String name;
-  @Column(name = "download_start_date_time")
-  private Date startDate;
 
-  @Column(name = "download_end_date_time")
-  private Date endDate;
+    @Id
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    private UUID id;
 
-  @Column(name = "total_elapsed_time")
-  private String elapsedTime;
-  @Column(name = "total_downloaded_kilobytes")
-  private String downloadedKilobytes;
+    private String website_name;
 
+    @JsonFormat(pattern="yyyy-MM-dd HH:mm:ss")
+    private LocalDateTime download_start_date_time;
 
-  public Website(){}
+    @JsonFormat(pattern="yyyy-MM-dd HH:mm:ss")
+    private LocalDateTime download_end_date_time;
 
-  public Website(String name, Date startDate, Date endDate, String elapsedTime, String downloadedKilobytes) {
-    this.name = name;
-    this.startDate = startDate;
-    this.endDate = endDate;
-    this.elapsedTime = elapsedTime;
-    this.downloadedKilobytes = downloadedKilobytes;
-  }
+    private Long  total_elapsed_time;
 
-  public UUID getId() {
-    return id;
-  }
+    private Long  total_downloaded_kilobytes;
 
-  public void setId(UUID id) {
-    this.id = id;
-  }
-
-  public String getName() {
-    return name;
-  }
-
-  public void setName(String name) {
-    this.name = name;
-  }
-
-  public Date getStartDate() {
-    return startDate;
-  }
-
-  public void setStartDate(Date startDate) {
-    this.startDate = startDate;
-  }
-
-  public Date getEndDate() {
-    return endDate;
-  }
-
-  public void setEndDate(Date endDate) {
-    this.endDate = endDate;
-  }
-
-  public String getElapsedTime() {
-    return elapsedTime;
-  }
-
-  public void setElapsedTime(String elapsedTime) {
-    this.elapsedTime = elapsedTime;
-  }
-
-  public String getDownloadedKilobytes() {
-    return downloadedKilobytes;
-  }
-
-  public void setDownloadedKilobytes(String downloadedKilobytes) {
-    this.downloadedKilobytes = downloadedKilobytes;
-  }
-
-  @Override
-  public String toString() {
-    return "Website{" +
-        "id=" + id +
-        ", website_name='" + name + '\'' +
-        ", start_time='" + startDate + '\'' +
-        ", end_time='" + endDate + '\'' +
-        ", elapsed_time='" + elapsedTime + '\'' +
-        ", total_kilobytes=" + downloadedKilobytes +
-        '}';
-  }
-
+    @JsonManagedReference
+    @OneToMany(mappedBy = "website",  cascade = CascadeType.REMOVE, orphanRemoval = true)
+    private Set<Link> links = new HashSet<>();
 }
